fix: update table when adding task
